phone-call-light/requirements.txt
flask>=3.0
RPi.GPIO>=0.7
